+++ Js2IL.Tests/CommonJS/JavaScript/CommonJS_Require_Reassigned_Function.js
@@ -1,0 +1,4 @@
+"use strict";
+
+require = (id) => 'custom:' + id;
+console.log(require('fs'));